use @zkochan/cmd-shim
bin shims have a lot of edge cases to consider, so instead
of re-implementing them we just reuse a library that does this.
This library is the same as the one used by yarn.

Using this library requires to bundle the output to still be able to
provide a standalone package.
When bundling, we cannot use require or require.resolve anymore for
resolving modules at runtime, so we need to manually resolve modules.